refactor(server): splitting  metadata parser code
